Removed sections page , replaced it with Home page
