+++ src/App.js
@@ -7,7 +7,6 @@ import LandingPage from './components/pages/landing-page/LandingPage';
 import SignIn from './components/pages/sign-in/SignIn';
 import SignUp from './components/pages/sign-up/SignUp';
 import Home from './components/pages/home/Home';
-import Sections from './components/pages/sections/Sections';
 import PageNotFound from './components/pages/PageNotFound';
 import UserProfilePage from './components/pages/UserProfilePage';
 
@@ -24,8 +23,7 @@ function App() {
               <Route path="/"  exact component={LandingPage}/>
               <div className="routes-with-navbar">
                 <Navbar />
-                <Route path="/home" exact component={Home}/>
-                <Route path="/sections" exact component={Sections}/> 
+                <Route path="/home" exact component={Home}/>              
                 <Route path="/user/:id" exact component={UserProfilePage}/>
               </div> 
             </Switch>
